backend/requirements.txt

# Various libraries
requests==2.31.0
bs4==0.0.1
joblib==1.3.1
bcrypt==4.1.2
python-dotenv==1.0.1
pandas==2.2.3


# Flask
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-Login==0.6.3


# Bitcoin specific
pyln-proto==23.5.2
jsonrpclib-pelix==0.4.3.2
bech32==1.2.0
python-bitcoinrpc==1.0


# Charts
matplotlib==3.8.3

# Tests
pytest==8.3.3

# Mysql
mysql-connector-python==9.0.0

# Prefect
prefect==3.0.8
